C_[이영석] 예약상세페이지구현 (전용우,한홍택) (#78)
* Combine jsp view controller

* Rolling modulize

* Image uploads page

* Reservation detail queries

* Get product images api temp

* Product title and bottom detail, location implement

* Comment list in detail page

* User comment image load

* Refactoring by mentoring

* Refactoring FE by mentoring

* Modify product main page url

* Flikcing.js modulize

* User comment photo viewr

* [BE] refactoring by mentoring

* [FE] refactoring by mentoring

* [BE] naver login

* [BE] change autowired from field to constructor

* [BE] product list image load

* [FE] naver login and product list image load

* [FE] Fix comment photo popup event on pc

* [FE] Implement click filicking event

* [FE] Remove scrollbar on comment photo popup

* [FE] Merge flicking and rolling to using prototype

* [FE] Use node_modules library

* [FE] Refactor variable name and ajax function

* [BE] Implement product reservation

* [FE] Implement product reservation

+++ reservation-service/src/main/java/com/ys/reservation/service/CategoryService.java
@@ -10,26 +10,35 @@ import com.ys.reservation.domain.Category;
 
 @Service //여기서 변수 검증해야함
 public class CategoryService {
-	
-	@Autowired
 	private CategoryDao categoryDao;
-	
+
+	@Autowired
+	public CategoryService(CategoryDao categoryDao) {
+		this.categoryDao = categoryDao;
+	}
+
 	public List<Category> getAll(){
 		return categoryDao.selectAll();
 	}
-	
+
 	public int create(Category category){
-		if(category.getName() != null && ! category.getName().trim().isEmpty()){ 
+		if(category != null && category.getName() != null && ! category.getName().trim().isEmpty()){ 
 			return categoryDao.insert(category);
 		}
 		return -1;
 	}
 	
 	public int remove(int id){
+		if(id < 1) {
+			return -1;
+		}
 		return categoryDao.delete(id);
 	}
 	
 	public int update(Category category){
+		if(category == null) {
+			return -1;
+		}
 		return categoryDao.update(category);
 	}
